refactor: renamed kapchat to twitch, reworked icons
