+++ i18n/locales/de-DE.json
@@ -31,7 +31,7 @@
     "settings": {
         "aside": {
             "1": "Allgemein",
-            "2": "KapChat",
+            "2": "Twitch",
             "3": "Eigenes CSS",
             "4": "Eigenes JavaScript"
         },
