+++ i18n/locales/en-US.json
@@ -36,7 +36,7 @@
     "settings": {
         "aside": {
             "1": "General",
-            "2": "KapChat",
+            "2": "Twitch",
             "3": "Custom CSS",
             "4": "Custom JavaScript"
         },
