+++ i18n/locales/es-AR.json
@@ -28,7 +28,7 @@
     "settings": {
         "aside": {
             "1": "General",
-            "2": "KapChat",
+            "2": "Twitch",
             "3": "CSS modificado",
             "4": "JavaScript modificado"
         },
