+++ i18n/locales/ko-KR.json
@@ -30,7 +30,7 @@
     "settings": {
         "aside": {
             "1": "일반",
-            "2": "KapChat",
+            "2": "Twitch",
             "3": "사용자 지정 CSS",
             "4": "사용자 지정 JavaScript"
         },
