+++ src/components/header/Buttons.vue
@@ -4,8 +4,6 @@ import { IpcEvent } from '@shared/constants';
 import { useRoute, useRouter } from 'vue-router/auto';
 import { ref, watch } from 'vue';
 
-defineEmits(['enableVanish']);
-
 const router = useRouter();
 const route = useRoute();
 
@@ -18,17 +16,17 @@ watch(route, () => {
 
 <template>
     <div id="menu-buttons">
-        <button v-if="showVanish" id="back" class="secondary" @click="$emit('enableVanish')">
-            <font-awesome-icon icon="fas fa-ghost" />
+        <button v-if="showVanish" id="back" class="secondary" @click="ipcRenderer.send(IpcEvent.Vanish)">
+            <font-awesome-icon icon="fa fa-ghost" />
         </button>
         <button id="back" class="secondary" @click="router.push('/')">
-            <font-awesome-icon icon="fas fa-chevron-left" />
+            <font-awesome-icon icon="fa fa-chevron-left" />
         </button>
         <button id="minimize" class="secondary" @click="ipcRenderer.send(IpcEvent.Minimize)">
-            <font-awesome-icon icon="fas fa-down-left-and-up-right-to-center" />
+            <font-awesome-icon icon="fa fa-down-left-and-up-right-to-center" />
         </button>
         <button id="close" class="secondary" @click="ipcRenderer.send(IpcEvent.Close)">
-            <font-awesome-icon icon="fas fa-xmark" />
+            <font-awesome-icon icon="fa fa-xmark" />
         </button>
     </div>
 </template>
