+++ vite.config.mts
@@ -18,6 +18,7 @@ export default defineConfig((({ command }) => {
     return {
         plugins: [
             VueRouter(),
+
             vue({
                 template: {
                     compilerOptions: {
@@ -25,6 +26,7 @@ export default defineConfig((({ command }) => {
                     },
                 },
             }),
+
             VueI18nPlugin.vite({
                 include: resolve(__dirname, './i18n/locales/**'),
                 runtimeOnly: false,
